Part of admin part. Buffered channel. Config part clean up.

+++ src/model/user.go
@@ -6,6 +6,7 @@ package model
 
 import (
     "OIUP-Backend/config"
+    "errors"
 )
 
 const (
@@ -23,24 +24,40 @@ type UserInfo struct {
 }
 
 func AddUser(user UserInfo) error {
+    _, found, err := GetUser(user.ContestID)
+    if err != nil {
+        return err
+    }
+    if found {
+        return errors.New("user with contest_id " + user.ContestID + " already exists")
+    }
+
     addUserQuery, _ := db.Prepare("INSERT INTO " + config.Config.DB.TableUser +
         " VALUES (?, ?, ?, ?, ?)")
-    writeCh <- DBWriteQuery{
+    queryCh <- DBWriteQuery{
         Stmt:       addUserQuery,
         Parameters: []interface{}{user.Name, user.School, user.ContestID, user.PersonID, user.Language},
     }
-    err := <-errCh
+    err = <-errCh
     return err
 }
 
 func DeleteUser(contestID string) error {
+    _, found, err := GetUser(contestID)
+    if err != nil {
+        return err
+    }
+    if !found {
+        return errors.New("user with contest_id " + contestID + " not found")
+    }
+
     deleteUserQuery, _ := db.Prepare("DELETE FROM " + config.Config.DB.TableUser +
         " WHERE contest_id = ?")
-    writeCh <- DBWriteQuery{
+    queryCh <- DBWriteQuery{
         Stmt:       deleteUserQuery,
         Parameters: []interface{}{contestID},
     }
-    err := <-errCh
+    err = <-errCh
     return err
 }
 
@@ -63,12 +80,20 @@ func GetUser(contestID string) (UserInfo, bool, error) {
 }
 
 func UpdateUser(user UserInfo) error {
+    _, found, err := GetUser(user.ContestID)
+    if err != nil {
+        return err
+    }
+    if found {
+        return errors.New("user with contest_id " + user.ContestID + " already exists")
+    }
+
     updateUserQuery, _ := db.Prepare("UPDATE "  + config.Config.DB.TableUser +
         " SET name = ?, school = ?, person_id = ?, language = ? WHERE contest_id = ?")
-    writeCh <- DBWriteQuery{
+    queryCh <- DBWriteQuery{
         Stmt:       updateUserQuery,
         Parameters: []interface{}{user.Name, user.School, user.PersonID, user.Language, user.ContestID},
     }
-    err := <-errCh
+    err = <-errCh
     return err
 }
